SNES ex70.c: compare iteration counts to custom SIMPLE
From [petsc-maint #155997]:

When I compare my version of SIMPLE with the out-of-the-box
version, I see that the latter requires too many iterations for
the fieldsplit_1 solve because it only uses the jacobi
preconditioner, see below. In order to reproduce the results from
Elman e.a. Taxonomy paper, we will need to
replace "-fieldsplit_1_pc_type jacobi" by something more powerful
such as "-fieldsplit_1_pc_type gamg".

In an attempt to quantify the behaviour I've added a routine
"iterationCount" which was meant to give the total number of iterations
of the "fieldsplit_0_ksp" and "fieldsplit_1_ksp" but it only gives the
number for the last coupled iteration.

my version of SIMPLE:
-----------------------------
$ mpiexec -n 2 ./ex70 -nx 32 -ny 48 -ksp_type fgmres -pc_type fieldsplit -pc_fieldsplit_type schur -pc_fieldsplit_schur_fact_type lower -user_ksp
 iterations coupled system = 57
 iterations sub-system 0   = 1
 iterations sub-system 1   = 48
 residual u = 3.03871e-05
 residual p = 1.4044e-07
 residual [u,p] = 3.03874e-05
 discretization error u = 0.000183699
 discretization error p = 0.249002
 discretization error [u,p] = 0.249002

Out-of-the box version of SIMPLE:
--------------------------------------------
$ mpiexec -n 2 ./ex70 -nx 32 -ny 48 -ksp_type fgmres -pc_type fieldsplit -pc_fieldsplit_type schur -pc_fieldsplit_schur_fact_type lower -fieldsplit_0_ksp_type gmres -fieldsplit_0_pc_type bjacobi -fieldsplit_1_pc_type jacobi -fieldsplit_1_inner_ksp_type preonly -fieldsplit_1_inner_pc_type jacobi -fieldsplit_1_upper_ksp_type preonly -fieldsplit_1_upper_pc_type jacobi
 iterations coupled system = 43
 iterations sub-system 0   = 1
 iterations sub-system 1   = 351
 residual u = 3.20371e-05
 residual p = 1.5949e-07
 residual [u,p] = 3.20375e-05
 discretization error u = 0.000184736
 discretization error p = 0.248922
 discretization error [u,p] = 0.248922