PETSC_EXTERN in definitions for all interfaces involving function pointers
C++98 and C++11 specify in Section 7.5.1:

    Two function types with different language linkages are distinct
    types even if they are otherwise identical.

Most C++ compilers, including GCC, Clang, Intel, MSVC, IBM XL, PGI,
PathScale, and Open64, are in direct violation of this standard, but
they are in no rush to fix it since that would break lots of existing
code and implies an ABI change. See discussion here:

  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=2316

which contains comments such as

  "Finding reasons to ask for the removal of this feature from the next
  standard is kind of relevant ;-)"

  "And we may even decide on an official WONTFIX."

In any case, Cray and Sun compilers comply with Section 7.5.1, causing
code like the following to overload at the definition

  extern "C" int Frotz(int (*)(void)); // in a header
  int Frotz(int (*fptr)(void)) { ... } // in implementation

leading to the C symbol `Frotz` being undefined at link time.

This patch adds the extern "C" statement to all such definitions, so
that they do not overload. Note that we still do not actually keep
function pointers with C versus C++ language linkage separate, which is
formally undefined behavior, but *all* compilers we are aware of use
identical calling conventions for C versus C++.