TS ex33: start segmental refinement example
This is a simplified version of ts/ex11 that solves a Lapacian in 2D
with a 5-point stencil.  It has a synthetic solution and a convergence
test. It uses SNES to construct a differencing Jacoabian without a
solve.  This is intened to be the start of an extreme scale solver
driver.