Added dm/examples/tutorials/ex13f90
This new example uses DMDA to solve a toy problem which looks like a PDE. It is
intended to be a good "first introduction" to DMDA. It is split into two files,
where one contains a module with all subroutines and functions. In Fortran, this
ensures that we automatically have explicit interfaces so that type checking is
performed.