Plex: First implementation of DMPlexRedistribute()
This should do the same as DMPlexDistribute, but starting from an already
distributed Plex. The initial version will only be able to re-size the
overlap region; full parallel re-partitioning is the next step.

This commit implements the re-building of the partitioning
based on the PointSF, given that the old overlap spans the new one.