Add make warning when we suspect that next is upstream
Because some contributors
(http://lists.mcs.anl.gov/pipermail/petsc-dev/2014-September/015899.html)
will start their development on next, no matter how many exclamation
points surround the phrase **!! NEVER create a branch from `next` !!**
on the wiki.  The premise here is that the most due diligence we can
expect a contributor to perform is to check compilation of their
changes with `make`.  We therefore create a commit (not this commit that
you're reading) that only touches the file `.petscnextupstream`, and
only merge that commit into next (this commit should also be the first
merged into next anytime next is rewound).  In this commit, we add a
hook to `make info` (which is called by `make all`) that checks for the
presence of `.petscnextupstream`.  If this file is present, and the
current branch is not `next`, then either the current branch was based
on `next` or had `next` merged into it.  In that case, we print a
warning, so that we can hopefully reach contributors with even the
poorest of reading comprehension, and catch bad developments before they
start.