ViennaCL: Added pipelined CG as a preconditioner.
Similar to bicgstabcusp. Enable via
 -ksp_type preonly -pc_type pipecgviennacl
on the command line, or directly in code via PCPIPECGVIENNACL.
Additional options:
 -pc_pipecgviennacl_rtol ... rel. tolerance
 -pc_pipecgviennacl_max_it ... max number of iterations
 -pc_pipecgviennacl_monitor_verbose ... Print no. of iterations

Making such GPU-optimized KSPs available as preconditioners looks odd
at first sight, but emphasizes that this is more of a monolithic block
and not very composable. Some more thought is required on how to
support other KSPs from external packages in the future.