BuildSystem: fix framework.parallelQueueEvaluation() when loading framework
The method was working before because none of the incoming children had
the _configured attribute.  When they do (when loading a previous
framework), we need to be more careful.

Looking at the serial queue evaluation, even if a child vertex has the
_configured flag, something is still done (no_configure()), so we have
to divorce the decision to put something in the queue from that flag,
hence the new _queued flag.

I also put in a counter with a lock so that the worker threads know the
difference between the queue being temporarily empty and all work being
completed.