BuildSystem: introduce locks for contraflow state changes
Once configured, a module is not immutable: its children sometimes
change its state.  I don't know, maybe this means that the DAG is not
properly factored.  But if this type of behavior is going to be allowed,
then we need some way of avoiding race conditions when a child changes a
parent.  A lock was simple enough to write, but I by no means think that
I have caught all of the places that it should be applied, and for all I
know I may have introduced a hang, but it seems to work for me.  This
needs to be more rigorously tested.