BuildSystem: try __getattribute__ instead of __getattr__
We want a thread to be able to temporarily mask an attribute with a
thread-local version, without that change being visible to other
threads.  But `__getattr__()` only triggers when `self.X` cannot be
obtained by the normal method (e.g., `self.__dict__[X]`).  I tried to
make this work by moving  self.X to self._masked[X], but this got really
complicated, and would have required implementing `__setattr__()` as
well, and a plethora of locks would have to be used.  So I implemented
`__getattribute__()` which is called for *every* attribute access, but
does not require any locks.

working?