Plex: added DMPlexVecGet/RestoreStar()
This is to be used in the type of access patterns for double-valued jump
calculations on facets in FV and DG calculations, as well as explicit
posterior error estimators.  Right now, FV access patterns are supported
via hybrid/ghost cells.  I've never been happy with this, because the
topology of the plex changes depending on whether a finite volume method
is being used, which feels like a bad dependency inversion.  I think
Matt chose to do this so that facet calculations can be done quickly
using DMPlexGetSupport() on facets, with the expectation that the
support size is always 2.  Instead, if performance is critical I think
that a star section can be stashed in a PetscSection, in the same way
that a closure section can be stashed now.