MatNest support in DMNetwork
This is a new re-work of the previous pull request where support for
MatNest was implemented to allow efficient use of FieldSplit.

I'm only looking for a review or comments on the code. It is not ready to
merge.

In this code, an alternate ordering for the subsets of edges and vertices
is created in order to index submatrices.

This is accomplished by creating subsets of the petscSF that plex owns.
Local DoF sections are created and, using the 'subsf', Global DoF
sections are assembled reflecting the new order for the subset.

Ideally, accessing a MatNest type matrix should not require much
modification from the DMNetwork users part. In that regard, my code is
still lacking. Also, I would need to create sections for the data too.

The user might like to define submatrices for different sets outside the
vertex/edge separation. For example: two groups of edges having very
different physics. I believe the approach of creating 'subSF's is
flexible enought that could be adapted for this matter. But let me know
what do you think...

Thanks