Updates to examplesWalker/New scripts for convert
examplesWalker.py has been refactored substantially to be a class to
more easily allow the functionality for converting examples to be moved
over into a new script: convertExamples.py

convertExamples.py uses a convertExamplesUtils.py to define some of the
metadata language that gets inserted into the source files.  See
documentation at the top of that file for the preliminary specification.

It is useful to say what this does not do, or what remains to be done:
 -- It generates files of the form new_ex1.c; i.e., it does not replace
     the files yet.
 -- This commit is still about summarizing the files.  I have not
     started the slog towards cleaning up the makefiles by hand.
     The problems from the examplesWalker.py analysis still remains.
 -- Related to the above, I have not looked too much at further
     abstractions ideas; e.g, having a functionality for adding grep of
     output as a standard test.
 -- This commit does not include the scripts for going from source files
      to shell script (the inverse of convertExamples.py)