PetscQuadrature: introduce field dimension
We are using quadrature to also stand for general functionals.  For
non-Lagrange finite elements, the dual spaces may be vector valued, so
we need to have vector valued functionals.

I just changed the interface and put in some assertion guards for now to
ensure we are only using scalar quadrature.  Later I will go and look
closely at the usage to see where we need to allow for vector-valued
functionals.

WIP