convertExamples.py puts out into new languages
This is a start on getting all of the examples converted over
to the new language specified in the documentation.  To simplify
the development, this is cleaning up the code to make all of the
makefile specific (or directory specific) code into one file:
examplesMkParse.py.  Right now this runs and gives some preliminary
conversion, but more methods from examplesWalker need to be pushed down
to examplesMkParse.py and then allow examplesWalker.py to inherit from
that.