TSGradient: evaluate point functionals and their derivative wrt the parameters with the correct state vector
Before, the code was using the solution at the end of the step and a time in between the previous and current time.
Taylor remainder tests are now all fine.
Added a test case to ex23.c with parameter dependent objective function
Fix some compilation warning
Change golden files as some of them got changed after the previous commit