TS: added TSComputeHessian()
The code has been greatly refactored to support the Hessian matvec
Testes with ex16opt_p, ex16opt_ic and ex20opt_p.
Added matrix free runs for these examples
ex23 Hessians also works, except when the point-form functional is in between the simulation
This has to do with the way the trajectory of the first-order adjoint is stored. I need to think at a fix
Once this will be clean, I'll commit the runs for ex23