PetscDTAltv: New basis for alterting multilinear forms.
For $k$-forms with $2 k \leq N$, the standard lexicographic
basis of $\{dx_{\sigma(1)} \wedge \dots \wedge dx_{\sigma(k)}\}$
is used.

For $2 k > N$, we modify this by first reversing the lexicographic
order, and then swapping $\sigma(1)$ and $\sigma(k)$ if $\sigma$ is odd.
While this involves more sign flips than the basic approach, it has two
good properties:

- it preserves the case as before that the dofs of a row vector /
  covector are the dofs of the equivalent 1-form, but now
- the dofs of a contravariant vector (like an H(div)-conforming flux)
  are the dofs of the equivalent (N-1)-form (importantly, this means
  that the contravariant pullback is the same as the alternating form
  pullback)

The only exception is 2D, for which the case is the same as before: the
1-form pullback is the covariant pullback, the contravariant pullback
can be obtained by Hodge star isometry before and afer the pullback.