DMForest: fix nans in coordinate array
The design decision early on was that coarse-to-fine
would be a local operation that could be phrased
as a local matrix applied the PetscSFBcast in DMGlobalToLocal.

(I'm not sure I would make the same decision if I were starting now.)

A consequence is that there are mesh points in the local meshes
that are not in the closure of any cells.

Long story short, that means that the p4est routine that figures out the
coordinates (in a loop over cells) doesn't compute all coordinates if we
have overlapping cells.  An extra broadcast of owned coordinates is
needed to fill in the gaps.